Add multi-network training and UI
Added training logic for networks. Networks are trained using a genetic
algorithm with the game score as the fitness function.

Added basic UI and graphing to see the improvment in the networks over
time. At the moment something is wrong with my genetic algorithm which is
causing them to not improve at all. I believe the selection is too random.