Add animations back in
Animations for movement and merging have been re-inputted.

The input is handled in a different way. Instead of applying it straight
away, the lastest input request is stored, and applied when possible
(after the animations have finished).

The move animation cannot be interupted, but the merge animation can be.
This means the fastest an input can be handled is once every 0.1 seconds.